Use a connection pool for Postgres
It is not allowed to run concurrent queries on a single DB connection. We could solve this by:
1. Wrap the connection into a Mutex (but then we will need to wait to lock() the Mutex)
2. Use a pool of connections

The pool of connections will handle passing us a connection. When the request needs a connection, the pool will either create a new one (and give us exclusive access) or wait for one to be available.
This way we pay the cost of *establishing* a connection just when needed. (Amortized it's better)