Issue warning if any email is invalid
We have the invariant that all SubscriberEmails are valid, but the perception of "valid" may change between our application versions. In the case where an email was valid in the past version but is not valid in the current version, we emit a warning so that the operator can fix the problem, but otherwise continue delivering the newsletter for valid emails.