Generate an idempotency key for each request
If the request was already fulfilled (we have the same key stored), we just return the original HTTP response.
This works as expected if the "retry" request comes *after* the original request has been processed. We still need to work on what to do if the retry comes *during* the original request processing.